pyproject.toml
[project]
name = "wikipedia-rag"
version = "0.1.0"
description = "Wikipedia RAG System using ChromaDB and Google Gemini API"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "chromadb>=0.4.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
